requirements.txt
asgiref==3.7.2
asttokens==2.4.1
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dj-database-url==2.2.0
Django==4.2.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.12.4
Flask==3.0.0
gunicorn==22.0.0
idna==3.7
ipykernel==6.27.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
legacy==0.1.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysqlclient==2.2.4
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pillow==10.4.0
platformdirs==3.11.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.8.2
python-decouple==3.8
pywin32==306
pyzmq==25.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.12
Werkzeug==3.0.1
whitenoise==6.6.0
